% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getOpt.R
\name{getOpt}
\alias{getOpt}
\alias{setOpt}
\title{Manipulate the options}
\usage{
getOpt(opt = c("a", "b"))

setOpt(a, b)
}
\arguments{
\item{opt}{a possible value indicating the option(s) to return}

\item{a}{the value for option a}

\item{b}{the value for option b}
}
\value{
in case of \code{setOpt} the old options. In case of
\code{getOpt}, the options.
}
\description{
functions to manipulate the options.
}
\examples{
getOpt()
getOpt("b")

setOpt( a = "another value")
getOpt("a")

}
